Add file attachments to posts in gencontroller
