+++ loadtest/control/utils.go
@@ -13,9 +13,13 @@ import (
 	"regexp"
 	"runtime"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/blang/semver"
+	"github.com/mattermost/mattermost-load-test-ng/loadtest/user"
+	"github.com/mattermost/mattermost-server/server/v8/model"
+	"github.com/wiggin77/merror"
 )
 
 type PostsSearchOpts struct {
@@ -260,4 +264,61 @@ func IsVersionSupported(version, serverVersionString string) (bool, error) {
 	}
 
 	return v.LTE(sv), nil
+}
+
+// AttachFilesToPost uploads at least one file on behalf of the user, attaching
+// all uploaded files to the post.
+func AttachFilesToPost(u user.User, post *model.Post) error {
+	type file struct {
+		data   []byte
+		upload bool
+	}
+	filenames := []string{"test_upload.png", "test_upload.jpg", "test_upload.mp4"}
+	files := make(map[string]*file, len(filenames))
+
+	for _, filename := range filenames {
+		files[filename] = &file{
+			data:   MustAsset(filename),
+			upload: rand.Intn(2) == 0,
+		}
+	}
+
+	// We make sure at least one file gets uploaded.
+	files[filenames[rand.Intn(len(filenames))]].upload = true
+
+	var wg sync.WaitGroup
+	fileIdsChan := make(chan string, len(files))
+	errChan := make(chan error, len(files))
+	for filename, file := range files {
+		if !file.upload {
+			continue
+		}
+		wg.Add(1)
+		go func(filename string, data []byte) {
+			defer wg.Done()
+			resp, err := u.UploadFile(data, post.ChannelId, filename)
+			if err != nil {
+				errChan <- err
+				return
+			}
+			fileIdsChan <- resp.FileInfos[0].Id
+		}(filename, file.data)
+	}
+
+	wg.Wait()
+	close(fileIdsChan)
+	close(errChan)
+
+	// Attach all successfully uploaded files
+	for fileId := range fileIdsChan {
+		post.FileIds = append(post.FileIds, fileId)
+	}
+
+	// Collect all errors
+	merr := merror.New()
+	for err := range errChan {
+		merr.Append(err)
+	}
+
+	return merr.ErrorOrNil()
 }
